+++ changelog.md
@@ -5,6 +5,7 @@
 - Fix "force hide cooldown numbers" option in the PlayerInfo module not applying to weapon enchants until a reload.
 - Fix Configuration Mode not showing placeholder buffs and debuffs on Info modules on 12.1+ clients.
 - Fix Info module buffs and debuffs not always respecting the configured anchors and grow directions correctly on 12.1+ clients.
+- Fix cooldown wipes on Info modules to not be fully opaque.
 
 ## v1.17.24